Refactor: simplify run script
