+++ pyproject.toml
@@ -31,7 +31,7 @@ dev = [
 ]
 
 [project.scripts]
-oresat-gps = "oresat_gps.__main__:main"
+oresat-gps = "oresat_gps:main"
 
 [tool.setuptools.packages.find]
 exclude = ["docs*"]
